Backend/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.10.0
apispec==6.8.2
apispec-webframeworks==1.2.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
greenlet==3.2.4
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==4.0.0
mistune==3.1.3
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymssql==2.3.7
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
